split out listener parsing
