+++ CMakeLists.txt
@@ -2667,6 +2667,7 @@ add_library(grpc
   src/core/xds/grpc/xds_http_stateful_session_filter.cc
   src/core/xds/grpc/xds_lb_policy_registry.cc
   src/core/xds/grpc/xds_listener.cc
+  src/core/xds/grpc/xds_listener_parser.cc
   src/core/xds/grpc/xds_route_config.cc
   src/core/xds/grpc/xds_route_config_parser.cc
   src/core/xds/grpc/xds_routing.cc
