+++ Package.swift
@@ -1998,6 +1998,8 @@ let package = Package(
         "src/core/xds/grpc/xds_lb_policy_registry.h",
         "src/core/xds/grpc/xds_listener.cc",
         "src/core/xds/grpc/xds_listener.h",
+        "src/core/xds/grpc/xds_listener_parser.cc",
+        "src/core/xds/grpc/xds_listener_parser.h",
         "src/core/xds/grpc/xds_route_config.cc",
         "src/core/xds/grpc/xds_route_config.h",
         "src/core/xds/grpc/xds_route_config_parser.cc",
