+++ build_autogenerated.yaml
@@ -1239,6 +1239,7 @@ libs:
   - src/core/xds/grpc/xds_http_stateful_session_filter.h
   - src/core/xds/grpc/xds_lb_policy_registry.h
   - src/core/xds/grpc/xds_listener.h
+  - src/core/xds/grpc/xds_listener_parser.h
   - src/core/xds/grpc/xds_route_config.h
   - src/core/xds/grpc/xds_route_config_parser.h
   - src/core/xds/grpc/xds_routing.h
@@ -2046,6 +2047,7 @@ libs:
   - src/core/xds/grpc/xds_http_stateful_session_filter.cc
   - src/core/xds/grpc/xds_lb_policy_registry.cc
   - src/core/xds/grpc/xds_listener.cc
+  - src/core/xds/grpc/xds_listener_parser.cc
   - src/core/xds/grpc/xds_route_config.cc
   - src/core/xds/grpc/xds_route_config_parser.cc
   - src/core/xds/grpc/xds_routing.cc
