+++ gRPC-C++.podspec
@@ -1348,6 +1348,7 @@ Pod::Spec.new do |s|
                       'src/core/xds/grpc/xds_http_stateful_session_filter.h',
                       'src/core/xds/grpc/xds_lb_policy_registry.h',
                       'src/core/xds/grpc/xds_listener.h',
+                      'src/core/xds/grpc/xds_listener_parser.h',
                       'src/core/xds/grpc/xds_route_config.h',
                       'src/core/xds/grpc/xds_route_config_parser.h',
                       'src/core/xds/grpc/xds_routing.h',
@@ -2628,6 +2629,7 @@ Pod::Spec.new do |s|
                               'src/core/xds/grpc/xds_http_stateful_session_filter.h',
                               'src/core/xds/grpc/xds_lb_policy_registry.h',
                               'src/core/xds/grpc/xds_listener.h',
+                              'src/core/xds/grpc/xds_listener_parser.h',
                               'src/core/xds/grpc/xds_route_config.h',
                               'src/core/xds/grpc/xds_route_config_parser.h',
                               'src/core/xds/grpc/xds_routing.h',
