+++ gRPC-Core.podspec
@@ -2113,6 +2113,8 @@ Pod::Spec.new do |s|
                       'src/core/xds/grpc/xds_lb_policy_registry.h',
                       'src/core/xds/grpc/xds_listener.cc',
                       'src/core/xds/grpc/xds_listener.h',
+                      'src/core/xds/grpc/xds_listener_parser.cc',
+                      'src/core/xds/grpc/xds_listener_parser.h',
                       'src/core/xds/grpc/xds_route_config.cc',
                       'src/core/xds/grpc/xds_route_config.h',
                       'src/core/xds/grpc/xds_route_config_parser.cc',
@@ -3404,6 +3406,7 @@ Pod::Spec.new do |s|
                               'src/core/xds/grpc/xds_http_stateful_session_filter.h',
                               'src/core/xds/grpc/xds_lb_policy_registry.h',
                               'src/core/xds/grpc/xds_listener.h',
+                              'src/core/xds/grpc/xds_listener_parser.h',
                               'src/core/xds/grpc/xds_route_config.h',
                               'src/core/xds/grpc/xds_route_config_parser.h',
                               'src/core/xds/grpc/xds_routing.h',
