+++ package.xml
@@ -1982,6 +1982,8 @@
     <file baseinstalldir="/" name="src/core/xds/grpc/xds_lb_policy_registry.h" role="src" />
     <file baseinstalldir="/" name="src/core/xds/grpc/xds_listener.cc" role="src" />
     <file baseinstalldir="/" name="src/core/xds/grpc/xds_listener.h" role="src" />
+    <file baseinstalldir="/" name="src/core/xds/grpc/xds_listener_parser.cc" role="src" />
+    <file baseinstalldir="/" name="src/core/xds/grpc/xds_listener_parser.h" role="src" />
     <file baseinstalldir="/" name="src/core/xds/grpc/xds_route_config.cc" role="src" />
     <file baseinstalldir="/" name="src/core/xds/grpc/xds_route_config.h" role="src" />
     <file baseinstalldir="/" name="src/core/xds/grpc/xds_route_config_parser.cc" role="src" />
