+++ src/core/resolver/xds/xds_dependency_manager.cc
@@ -28,6 +28,7 @@
 #include "src/core/load_balancing/xds/xds_channel_args.h"
 #include "src/core/resolver/fake/fake_resolver.h"
 #include "src/core/xds/grpc/xds_route_config_parser.h"
+#include "src/core/xds/grpc/xds_listener_parser.h"
 #include "src/core/xds/grpc/xds_routing.h"
 
 namespace grpc_core {
