+++ src/python/grpcio/grpc_core_dependencies.py
@@ -851,6 +851,7 @@ CORE_SOURCE_FILES = [
     'src/core/xds/grpc/xds_http_stateful_session_filter.cc',
     'src/core/xds/grpc/xds_lb_policy_registry.cc',
     'src/core/xds/grpc/xds_listener.cc',
+    'src/core/xds/grpc/xds_listener_parser.cc',
     'src/core/xds/grpc/xds_route_config.cc',
     'src/core/xds/grpc/xds_route_config_parser.cc',
     'src/core/xds/grpc/xds_routing.cc',
