+++ tools/doxygen/Doxyfile.core.internal
@@ -2784,6 +2784,8 @@ src/core/xds/grpc/xds_lb_policy_registry.cc \
 src/core/xds/grpc/xds_lb_policy_registry.h \
 src/core/xds/grpc/xds_listener.cc \
 src/core/xds/grpc/xds_listener.h \
+src/core/xds/grpc/xds_listener_parser.cc \
+src/core/xds/grpc/xds_listener_parser.h \
 src/core/xds/grpc/xds_route_config.cc \
 src/core/xds/grpc/xds_route_config.h \
 src/core/xds/grpc/xds_route_config_parser.cc \
